starships-app/requirements.txt
chalice==1.26.0
PyJWT==2.9.0
requests
PyYAML==5.3.1